Refactor ImplicitSkinning part 1
